Confine formatting-related hooks to meaningful stages
Prevent these hooks to be default-installed for stages that either do not
make sense or would run the hooks on the files holding commit-messages & Co.
The later can lead to problems as git automatically inserts comments that
include tabs into the commit message for certain operations, such as
amending a commit.