Feature: Dynamic years with corresponding tests
Add flag '--dynamic-years' which determines the start year of the copyright time range automatically - based on when
the file was first tracked with Git. If a start year is already present, it is not touched.
If a file is not tracked by Git, the current year is used as start year.
The end year is automatically set to the current year
(--use-current-year is activated automatically when --dynamic-years is present).